+++ internal/generator/mutator/discrete.go
@@ -37,6 +37,20 @@ func generateForDiscreteField(code *jen.File, f *scope.Field) {
 				Id(f.Message.GoTypeName),
 		).
 		Block(
+			jen.If(
+				jen.
+					Id("x").
+					Op("==").
+					Nil(),
+			).
+				Block(
+					jen.
+						Id("x").
+						Op("=").
+						Op("&").
+						Id(f.Message.GoTypeName).
+						Values(),
+				),
 			jen.
 				Id("x").
 				Dot(f.GoFieldName).