+++ internal/test/mutators/helper_test.go
@@ -46,23 +46,32 @@ func testMutatorFunc[M message, T any](
 		panic("cannot test mutator with zero value")
 	}
 
-	var m M
-	m = reflect.New(
-		reflect.TypeOf(m).Elem(),
-	).Interface().(M)
+	var nil_ M
 
-	p := mutator(m, want)
-	if p != m {
+	t.Log("calling mutator on a nil message")
+	m := mutator(nil_, want)
+	if m == nil_ {
+		t.Fatal("mutator returned a nil message")
+	}
+
+	t.Log("checking that the mutator set the field")
+	if got := accessor(m); !eq(got, want) {
 		t.Fatalf(
-			"mutator did not return the message: got: %v, want: %v",
-			p,
-			m,
+			"mutator did not set the field: got: %v, want: %v",
+			got,
+			want,
 		)
 	}
 
-	got := accessor(m)
+	t.Log("calling mutator on a non-nil message")
+	var zero T
+	want = zero
+	r := mutator(m, want)
+	if r != m {
+		t.Fatal("mutator did not return the mutated message")
+	}
 
-	if !eq(got, want) {
+	if got := accessor(m); !eq(got, want) {
 		t.Fatalf(
 			"mutator did not set the field: got: %v, want: %v",
 			got,